python-services/backtrader-analysis/requirements.txt
# Core Trading and Analysis
backtrader==1.9.78.123
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
plotly==5.15.0
seaborn==0.12.2

# API Framework
flask==2.3.2
flask-cors==4.0.0
flask-restx==1.1.0
gunicorn==21.2.0

# Data Processing and Statistics
scipy==1.11.1
scikit-learn==1.3.0
statsmodels==0.14.0
ta-lib==0.4.26

# Utilities
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.1.1
click==8.1.6

# Development and Testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.4.1

# Logging and Monitoring
structlog==23.1.0
prometheus-client==0.17.1

# Image Processing for Charts
Pillow==10.0.0
kaleido==0.2.1

# Date and Time Utilities
python-dateutil==2.8.2
pytz==2023.3